+++ src/snowflake/snowpark/modin/plugin/_internal/aggregation_utils.py
@@ -8,7 +8,9 @@ import functools
 from collections import defaultdict
 from collections.abc import Hashable, Iterable
 from functools import partial
-from typing import Any, Callable, Literal, NamedTuple, Optional, Union
+from inspect import getmembers
+from types import BuiltinFunctionType
+from typing import Any, Callable, Literal, Mapping, NamedTuple, Optional, Union
 
 import numpy as np
 from pandas._typing import AggFuncType, AggFuncTypeBase
@@ -78,6 +80,9 @@ from snowflake.snowpark.types import (
 )
 
 AGG_NAME_COL_LABEL = "AGG_FUNC_NAME"
+_NUMPY_FUNCTION_TO_NAME = {
+    function: name for name, function in getmembers(np) if callable(function)
+}
 
 
 def array_agg_keepna(
@@ -1127,11 +1132,75 @@ def using_named_aggregations_for_func(func: Any) -> bool:
     )
 
 
-def format_kwargs_for_error_message(kwargs: dict[Any, Any]) -> str:
+def repr_aggregate_function(agg_func: AggFuncType, agg_kwargs: Mapping) -> str:
     """
-    Helper method to format a kwargs dictionary for an error message.
+    Represent an aggregation function as a string.
 
-    Returns a string containing the keys + values of kwargs formatted like so:
-    "key1=value1, key2=value2, ..."
+    Use this function to represent aggregation functions in error message to
+    the user. This function will hide sensitive information, like axis labels or
+    names of callables, in the function description.
+
+    Args:
+        agg_func: AggFuncType
+            The aggregation function from the user. This may be a list-like or a
+            dictionary containing multiple aggregations.
+        agg_kwargs: Mapping
+            The keyword arguments for the aggregation function.
+
+    Returns:
+        str
+            The representation of the aggregation function.
     """
-    return ", ".join([f"{key}={value}" for key, value in kwargs.items()])
+    if using_named_aggregations_for_func(agg_func):
+        # New axis labels are sensitive, so replace them with "new_label."
+        # Existing axis labels are sensitive, so replace them with "label."
+        return ", ".join(
+            f"new_label=(label, {repr_aggregate_function(f, agg_kwargs)})"
+            for _, f in agg_kwargs.values()
+        )
+    if isinstance(agg_func, str):
+        # Strings functions represent names of pandas functions, e.g.
+        # "sum" means to aggregate with pandas.Series.sum. string function
+        # identifiers are not sensitive.
+        return repr(agg_func)
+    if is_dict_like(agg_func):
+        # axis labels in the dictionary keys are sensitive, so replace them with
+        # "label."
+        return (
+            "{"
+            + ", ".join(
+                f"label: {repr_aggregate_function(agg_func[key], agg_kwargs)}"
+                for key in agg_func.keys()
+            )
+            + "}"
+        )
+    if is_list_like(agg_func):
+        return f"[{', '.join(repr_aggregate_function(func, agg_kwargs) for func in agg_func)}]"
+    if isinstance(agg_func, BuiltinFunctionType):
+        return repr(agg_func)
+
+    # for built-in classes like `list`, return "list" as opposed to repr(list),
+    # i.e. <class 'list'>, which would be confusing because the user is using
+    # `list` as a callable in this context.
+    if agg_func is list:
+        return "list"
+    if agg_func is tuple:
+        return "tuple"
+    if agg_func is set:
+        return "set"
+    if agg_func is str:
+        return "str"
+
+    # Format numpy aggregations, e.g. np.argmin should become "np.argmin"
+    if agg_func in _NUMPY_FUNCTION_TO_NAME:
+        return f"np.{_NUMPY_FUNCTION_TO_NAME[agg_func]}"
+
+    # agg_func should be callable at this point. pandas error messages at this
+    # point are not consistent, so choose one style of error message.
+    if not callable(agg_func):
+        raise ValueError("aggregation function is not callable")
+
+    # Return a constant string instead of some kind of function name to avoid
+    # exposing sensitive user input in the NotImplemented error message and
+    # thus in telemetry.
+    return "Callable"